Revert "Redesign site"
