--- .vscode/launch.json
@@ -1,20 +1,0 @@
-{
-    // Use IntelliSense to learn about possible attributes.
-    // Hover to view descriptions of existing attributes.
-    // For more information, visit: https://go.microsoft.com/fwlink/?linkid=830387
-    "version": "0.2.0",
-    "configurations": [
-        {
-            "name": "Python: Flask",
-            "type": "python",
-            "request": "launch",
-            "module": "dogecoinisawesome",
-            "env": {
-                "FLASK_APP": "dogecoinisawesome.py",
-                "FLASK_ENV": "development"
-            },
-            "args": [],
-            "jinja": true
-        }
-    ]
-}
