+++ Procfile.windows
@@ -1,0 +1,1 @@
+web: py -m flask run
