--- app.json
@@ -1,26 +1,0 @@
-{
-  "name": "DogecoinIsAwesome.com",
-  "description": "Find all you need to know to get started with Dogecoin.",
-  "keywords": ["dogecoin", "is", "awesome"],
-  "website": "https://www.dogecoinisawesome.com/",
-  "repository": "https://github.com/dogecoinisawesome-dot-com/DogecoinIsAwesome.com",
-  "logo": "https://www.dogecoinisawesome.com/static/img/doge.png",
-  "success_url": "/",
-  "env": {
-    "FLASK_APP": {
-      "description": "Name of the module to import at flask run.",
-      "value": "dogecoinisawesome.py"
-    },
-    "FLASK_ENV": {
-      "description": "Setting FLASK_ENV to development will enable debug mode.",
-      "value": "development"
-    }
-  },
-  "formation": {
-    "web": {
-      "quantity": 1,
-      "size": "free"
-    }
-  },
-  "image": "heroku/python"
-}
