--- app/__init__.py
@@ -1,5 +1,0 @@
-from flask import Flask
-
-app = Flask(__name__, static_folder="static", static_url_path="")
-
-from app import routes
