--- app/static/js/webflow.js
@@ -1,29 +1,0 @@
-/*!
- * Webflow: Front-end site library
- * @license MIT
- * Inline scripts may access the api using an async handler:
- *   var Webflow = Webflow || [];
- *   Webflow.push(readyFunction);
- */!function(t){var e={};function n(i){if(e[i])return e[i].exports;var r=e[i]={i:i,l:!1,exports:{}};return t[i].call(r.exports,r,r.exports,n),r.l=!0,r.exports}n.m=t,n.c=e,n.d=function(t,e,i){n.o(t,e)||Object.defineProperty(t,e,{enumerable:!0,get:i})},n.r=function(t){"undefined"!=typeof Symbol&&Symbol.toStringTag&&Object.defineProperty(t,Symbol.toStringTag,{value:"Module"}),Object.defineProperty(t,"__esModule",{value:!0})},n.t=function(t,e){if(1&e&&(t=n(t)),8&e)return t;if(4&e&&"object"==typeof t&&t&&t.__esModule)return t;var i=Object.create(null);if(n.r(i),Object.defineProperty(i,"default",{enumerable:!0,value:t}),2&e&&"string"!=typeof t)for(var r in t)n.d(i,r,function(e){return t[e]}.bind(null,r));return i},n.n=function(t){var e=t&&t.__esModule?function(){return t.default}:function(){return t};return n.d(e,"a",e),e},n.o=function(t,e){return Object.prototype.hasOwnProperty.call(t,e)},n.p="",n(n.s=3)}([function(t,e,n){"use strict";var i={},r={},o=[],a=window.Webflow||[],s=window.jQuery,u=s(window),c=s(document),l=s.isFunction,f=i._=n(5),d=i.tram=n(1)&&s.tram,h=!1,p=!1;function v(t){i.env()&&(l(t.design)&&u.on("__wf_design",t.design),l(t.preview)&&u.on("__wf_preview",t.preview)),l(t.destroy)&&u.on("__wf_destroy",t.destroy),t.ready&&l(t.ready)&&function(t){if(h)return void t.ready();if(f.contains(o,t.ready))return;o.push(t.ready)}(t)}function m(t){l(t.design)&&u.off("__wf_design",t.design),l(t.preview)&&u.off("__wf_preview",t.preview),l(t.destroy)&&u.off("__wf_destroy",t.destroy),t.ready&&l(t.ready)&&function(t){o=f.filter(o,function(e){return e!==t.ready})}(t)}d.config.hideBackface=!1,d.config.keepInherited=!0,i.define=function(t,e,n){r[t]&&m(r[t]);var i=r[t]=e(s,f,n)||{};return v(i),i},i.require=function(t){return r[t]},i.push=function(t){h?l(t)&&t():a.push(t)},i.env=function(t){var e=window.__wf_design,n=void 0!==e;return t?"design"===t?n&&e:"preview"===t?n&&!e:"slug"===t?n&&window.__wf_slug:"editor"===t?window.WebflowEditor:"test"===t?window.__wf_test:"frame"===t?window!==window.top:void 0:n};var w,g=navigator.userAgent.toLowerCase(),b=i.env.touch="ontouchstart"in window||window.DocumentTouch&&document instanceof window.DocumentTouch,y=i.env.chrome=/chrome/.test(g)&&/Google/.test(navigator.vendor)&&parseInt(g.match(/chrome\/(\d+)\./)[1],10),x=i.env.ios=/(ipod|iphone|ipad)/.test(g);i.env.safari=/safari/.test(g)&&!y&&!x,b&&c.on("touchstart mousedown",function(t){w=t.target}),i.validClick=b?function(t){return t===w||s.contains(t,w)}:function(){return!0};var k,_="resize.webflow orientationchange.webflow load.webflow";function E(t,e){var n=[],i={};return i.up=f.throttle(function(t){f.each(n,function(e){e(t)})}),t&&e&&t.on(e,i.up),i.on=function(t){"function"==typeof t&&(f.contains(n,t)||n.push(t))},i.off=function(t){n=arguments.length?f.filter(n,function(e){return e!==t}):[]},i}function O(t){l(t)&&t()}function A(){k&&(k.reject(),u.off("load",k.resolve)),k=new s.Deferred,u.on("load",k.resolve)}i.resize=E(u,_),i.scroll=E(u,"scroll.webflow resize.webflow orientationchange.webflow load.webflow"),i.redraw=E(),i.location=function(t){window.location=t},i.env()&&(i.location=function(){}),i.ready=function(){h=!0,p?(p=!1,f.each(r,v)):f.each(o,O),f.each(a,O),i.resize.up()},i.load=function(t){k.then(t)},i.destroy=function(t){t=t||{},p=!0,u.triggerHandler("__wf_destroy"),null!=t.domready&&(h=t.domready),f.each(r,m),i.resize.off(),i.scroll.off(),i.redraw.off(),o=[],a=[],"pending"===k.state()&&A()},s(i.ready),A(),t.exports=window.Webflow=i},function(t,e,n){"use strict";var i=n(2)(n(6));window.tram=function(t){function e(t,e){return(new L.Bare).init(t,e)}function n(t){return t.replace(/[A-Z]/g,function(t){return"-"+t.toLowerCase()})}function r(t){var e=parseInt(t.slice(1),16);return[e>>16&255,e>>8&255,255&e]}function o(t,e,n){return"#"+(1<<24|t<<16|e<<8|n).toString(16).slice(1)}function a(){}function s(t,e,n){c("Units do not match ["+t+"]: "+e+", "+n)}function u(t,e,n){if(void 0!==e&&(n=e),void 0===t)return n;var i=n;return Q.test(t)||!J.test(t)?i=parseInt(t,10):J.test(t)&&(i=1e3*parseFloat(t)),0>i&&(i=0),i==i?i:n}function c(t){B.debug&&window&&window.console.warn(t)}var l=function(t,e,n){function r(t){return"object"==(0,i.default)(t)}function o(t){return"function"==typeof t}function a(){}return function i(s,u){function c(){var t=new l;return o(t.init)&&t.init.apply(t,arguments),t}function l(){}u===n&&(u=s,s=Object),c.Bare=l;var f,d=a[t]=s[t],h=l[t]=c[t]=new a;return h.constructor=c,c.mixin=function(e){return l[t]=c[t]=i(c,e)[t],c},c.open=function(t){if(f={},o(t)?f=t.call(c,h,d,c,s):r(t)&&(f=t),r(f))for(var n in f)e.call(f,n)&&(h[n]=f[n]);return o(h.init)||(h.init=s),c},c.open(u)}}("prototype",{}.hasOwnProperty),f={ease:["ease",function(t,e,n,i){var r=(t/=i)*t,o=r*t;return e+n*(-2.75*o*r+11*r*r+-15.5*o+8*r+.25*t)}],"ease-in":["ease-in",function(t,e,n,i){var r=(t/=i)*t,o=r*t;return e+n*(-1*o*r+3*r*r+-3*o+2*r)}],"ease-out":["ease-out",function(t,e,n,i){var r=(t/=i)*t,o=r*t;return e+n*(.3*o*r+-1.6*r*r+2.2*o+-1.8*r+1.9*t)}],"ease-in-out":["ease-in-out",function(t,e,n,i){var r=(t/=i)*t,o=r*t;return e+n*(2*o*r+-5*r*r+2*o+2*r)}],linear:["linear",function(t,e,n,i){return n*t/i+e}],"ease-in-quad":["cubic-bezier(0.550, 0.085, 0.680, 0.530)",function(t,e,n,i){return n*(t/=i)*t+e}],"ease-out-quad":["cubic-bezier(0.250, 0.460, 0.450, 0.940)",function(t,e,n,i){return-n*(t/=i)*(t-2)+e}],"ease-in-out-quad":["cubic-bezier(0.455, 0.030, 0.515, 0.955)",function(t,e,n,i){return(t/=i/2)<1?n/2*t*t+e:-n/2*(--t*(t-2)-1)+e}],"ease-in-cubic":["cubic-bezier(0.550, 0.055, 0.675, 0.190)",function(t,e,n,i){return n*(t/=i)*t*t+e}],"ease-out-cubic":["cubic-bezier(0.215, 0.610, 0.355, 1)",function(t,e,n,i){return n*((t=t/i-1)*t*t+1)+e}],"ease-in-out-cubic":["cubic-bezier(0.645, 0.045, 0.355, 1)",function(t,e,n,i){return(t/=i/2)<1?n/2*t*t*t+e:n/2*((t-=2)*t*t+2)+e}],"ease-in-quart":["cubic-bezier(0.895, 0.030, 0.685, 0.220)",function(t,e,n,i){return n*(t/=i)*t*t*t+e}],"ease-out-quart":["cubic-bezier(0.165, 0.840, 0.440, 1)",function(t,e,n,i){return-n*((t=t/i-1)*t*t*t-1)+e}],"ease-in-out-quart":["cubic-bezier(0.770, 0, 0.175, 1)",function(t,e,n,i){return(t/=i/2)<1?n/2*t*t*t*t+e:-n/2*((t-=2)*t*t*t-2)+e}],"ease-in-quint":["cubic-bezier(0.755, 0.050, 0.855, 0.060)",function(t,e,n,i){return n*(t/=i)*t*t*t*t+e}],"ease-out-quint":["cubic-bezier(0.230, 1, 0.320, 1)",function(t,e,n,i){return n*((t=t/i-1)*t*t*t*t+1)+e}],"ease-in-out-quint":["cubic-bezier(0.860, 0, 0.070, 1)",function(t,e,n,i){return(t/=i/2)<1?n/2*t*t*t*t*t+e:n/2*((t-=2)*t*t*t*t+2)+e}],"ease-in-sine":["cubic-bezier(0.470, 0, 0.745, 0.715)",function(t,e,n,i){return-n*Math.cos(t/i*(Math.PI/2))+n+e}],"ease-out-sine":["cubic-bezier(0.390, 0.575, 0.565, 1)",function(t,e,n,i){return n*Math.sin(t/i*(Math.PI/2))+e}],"ease-in-out-sine":["cubic-bezier(0.445, 0.050, 0.550, 0.950)",function(t,e,n,i){return-n/2*(Math.cos(Math.PI*t/i)-1)+e}],"ease-in-expo":["cubic-bezier(0.950, 0.050, 0.795, 0.035)",function(t,e,n,i){return 0===t?e:n*Math.pow(2,10*(t/i-1))+e}],"ease-out-expo":["cubic-bezier(0.190, 1, 0.220, 1)",function(t,e,n,i){return t===i?e+n:n*(1-Math.pow(2,-10*t/i))+e}],"ease-in-out-expo":["cubic-bezier(1, 0, 0, 1)",function(t,e,n,i){return 0===t?e:t===i?e+n:(t/=i/2)<1?n/2*Math.pow(2,10*(t-1))+e:n/2*(2-Math.pow(2,-10*--t))+e}],"ease-in-circ":["cubic-bezier(0.600, 0.040, 0.980, 0.335)",function(t,e,n,i){return-n*(Math.sqrt(1-(t/=i)*t)-1)+e}],"ease-out-circ":["cubic-bezier(0.075, 0.820, 0.165, 1)",function(t,e,n,i){return n*Math.sqrt(1-(t=t/i-1)*t)+e}],"ease-in-out-circ":["cubic-bezier(0.785, 0.135, 0.150, 0.860)",function(t,e,n,i){return(t/=i/2)<1?-n/2*(Math.sqrt(1-t*t)-1)+e:n/2*(Math.sqrt(1-(t-=2)*t)+1)+e}],"ease-in-back":["cubic-bezier(0.600, -0.280, 0.735, 0.045)",function(t,e,n,i,r){return void 0===r&&(r=1.70158),n*(t/=i)*t*((r+1)*t-r)+e}],"ease-out-back":["cubic-bezier(0.175, 0.885, 0.320, 1.275)",function(t,e,n,i,r){return void 0===r&&(r=1.70158),n*((t=t/i-1)*t*((r+1)*t+r)+1)+e}],"ease-in-out-back":["cubic-bezier(0.680, -0.550, 0.265, 1.550)",function(t,e,n,i,r){return void 0===r&&(r=1.70158),(t/=i/2)<1?n/2*t*t*((1+(r*=1.525))*t-r)+e:n/2*((t-=2)*t*((1+(r*=1.525))*t+r)+2)+e}]},d={"ease-in-back":"cubic-bezier(0.600, 0, 0.735, 0.045)","ease-out-back":"cubic-bezier(0.175, 0.885, 0.320, 1)","ease-in-out-back":"cubic-bezier(0.680, 0, 0.265, 1)"},h=document,p=window,v="bkwld-tram",m=/[\-\.0-9]/g,w=/[A-Z]/,g="number",b=/^(rgb|#)/,y=/(em|cm|mm|in|pt|pc|px)$/,x=/(em|cm|mm|in|pt|pc|px|%)$/,k=/(deg|rad|turn)$/,_="unitless",E=/(all|none) 0s ease 0s/,O=/^(width|height)$/,A=" ",S=h.createElement("a"),T=["Webkit","Moz","O","ms"],C=["-webkit-","-moz-","-o-","-ms-"],R=function(t){if(t in S.style)return{dom:t,css:t};var e,n,i="",r=t.split("-");for(e=0;e<r.length;e++)i+=r[e].charAt(0).toUpperCase()+r[e].slice(1);for(e=0;e<T.length;e++)if((n=T[e]+i)in S.style)return{dom:n,css:C[e]+t}},z=e.support={bind:Function.prototype.bind,transform:R("transform"),transition:R("transition"),backface:R("backface-visibility"),timing:R("transition-timing-function")};if(z.transition){var M=z.timing.dom;if(S.style[M]=f["ease-in-back"][0],!S.style[M])for(var D in d)f[D][0]=d[D]}var I=e.frame=function(){var t=p.requestAnimationFrame||p.webkitRequestAnimationFrame||p.mozRequestAnimationFrame||p.oRequestAnimationFrame||p.msRequestAnimationFrame;return t&&z.bind?t.bind(p):function(t){p.setTimeout(t,16)}}(),P=e.now=function(){var t=p.performance,e=t&&(t.now||t.webkitNow||t.msNow||t.mozNow);return e&&z.bind?e.bind(t):Date.now||function(){return+new Date}}(),j=l(function(e){function r(t,e){var n=function(t){for(var e=-1,n=t?t.length:0,i=[];++e<n;){var r=t[e];r&&i.push(r)}return i}((""+t).split(A)),i=n[0];e=e||{};var r=Y[i];if(!r)return c("Unsupported property: "+i);if(!e.weak||!this.props[i]){var o=r[0],a=this.props[i];return a||(a=this.props[i]=new o.Bare),a.init(this.$el,n,r,e),a}}function o(t,e,n){if(t){var o=(0,i.default)(t);if(e||(this.timer&&this.timer.destroy(),this.queue=[],this.active=!1),"number"==o&&e)return this.timer=new H({duration:t,context:this,complete:a}),void(this.active=!0);if("string"==o&&e){switch(t){case"hide":l.call(this);break;case"stop":s.call(this);break;case"redraw":f.call(this);break;default:r.call(this,t,n&&n[1])}return a.call(this)}if("function"==o)return void t.call(this,this);if("object"==o){var c=0;h.call(this,t,function(t,e){t.span>c&&(c=t.span),t.stop(),t.animate(e)},function(t){"wait"in t&&(c=u(t.wait,0))}),d.call(this),c>0&&(this.timer=new H({duration:c,context:this}),this.active=!0,e&&(this.timer.complete=a));var p=this,v=!1,m={};I(function(){h.call(p,t,function(t){t.active&&(v=!0,m[t.name]=t.nextStyle)}),v&&p.$el.css(m)})}}}function a(){if(this.timer&&this.timer.destroy(),this.active=!1,this.queue.length){var t=this.queue.shift();o.call(this,t.options,!0,t.args)}}function s(t){var e;this.timer&&this.timer.destroy(),this.queue=[],this.active=!1,"string"==typeof t?(e={})[t]=1:e="object"==(0,i.default)(t)&&null!=t?t:this.props,h.call(this,e,p),d.call(this)}function l(){s.call(this),this.el.style.display="none"}function f(){this.el.offsetHeight}function d(){var t,e,n=[];for(t in this.upstream&&n.push(this.upstream),this.props)(e=this.props[t]).active&&n.push(e.string);n=n.join(","),this.style!==n&&(this.style=n,this.el.style[z.transition.dom]=n)}function h(t,e,i){var o,a,s,u,c=e!==p,l={};for(o in t)s=t[o],o in K?(l.transform||(l.transform={}),l.transform[o]=s):(w.test(o)&&(o=n(o)),o in Y?l[o]=s:(u||(u={}),u[o]=s));for(o in l){if(s=l[o],!(a=this.props[o])){if(!c)continue;a=r.call(this,o)}e.call(this,a,s)}i&&u&&i.call(this,u)}function p(t){t.stop()}function m(t,e){t.set(e)}function g(t){this.$el.css(t)}function b(t,n){e[t]=function(){return this.children?function(t,e){var n,i=this.children.length;for(n=0;i>n;n++)t.apply(this.children[n],e);return this}.call(this,n,arguments):(this.el&&n.apply(this,arguments),this)}}e.init=function(e){if(this.$el=t(e),this.el=this.$el[0],this.props={},this.queue=[],this.style="",this.active=!1,B.keepInherited&&!B.fallback){var n=X(this.el,"transition");n&&!E.test(n)&&(this.upstream=n)}z.backface&&B.hideBackface&&G(this.el,z.backface.css,"hidden")},b("add",r),b("start",o),b("wait",function(t){t=u(t,0),this.active?this.queue.push({options:t}):(this.timer=new H({duration:t,context:this,complete:a}),this.active=!0)}),b("then",function(t){return this.active?(this.queue.push({options:t,args:arguments}),void(this.timer.complete=a)):c("No active transition timer. Use start() or wait() before then().")}),b("next",a),b("stop",s),b("set",function(t){s.call(this,t),h.call(this,t,m,g)}),b("show",function(t){"string"!=typeof t&&(t="block"),this.el.style.display=t}),b("hide",l),b("redraw",f),b("destroy",function(){s.call(this),t.removeData(this.el,v),this.$el=this.el=null})}),L=l(j,function(e){function n(e,n){var i=t.data(e,v)||t.data(e,v,new j.Bare);return i.el||i.init(e),n?i.start(n):i}e.init=function(e,i){var r=t(e);if(!r.length)return this;if(1===r.length)return n(r[0],i);var o=[];return r.each(function(t,e){o.push(n(e,i))}),this.children=o,this}}),N=l(function(t){function e(){var t=this.get();this.update("auto");var e=this.get();return this.update(t),e}function n(t){var e=/rgba?\((\d+),\s*(\d+),\s*(\d+)/.exec(t);return(e?o(e[1],e[2],e[3]):t).replace(/#(\w)(\w)(\w)$/,"#$1$1$2$2$3$3")}var r=500,a="ease",s=0;t.init=function(t,e,n,i){this.$el=t,this.el=t[0];var o=e[0];n[2]&&(o=n[2]),Z[o]&&(o=Z[o]),this.name=o,this.type=n[1],this.duration=u(e[1],this.duration,r),this.ease=function(t,e,n){return void 0!==e&&(n=e),t in f?t:n}(e[2],this.ease,a),this.delay=u(e[3],this.delay,s),this.span=this.duration+this.delay,this.active=!1,this.nextStyle=null,this.auto=O.test(this.name),this.unit=i.unit||this.unit||B.defaultUnit,this.angle=i.angle||this.angle||B.defaultAngle,B.fallback||i.fallback?this.animate=this.fallback:(this.animate=this.transition,this.string=this.name+A+this.duration+"ms"+("ease"!=this.ease?A+f[this.ease][0]:"")+(this.delay?A+this.delay+"ms":""))},t.set=function(t){t=this.convert(t,this.type),this.update(t),this.redraw()},t.transition=function(t){this.active=!0,t=this.convert(t,this.type),this.auto&&("auto"==this.el.style[this.name]&&(this.update(this.get()),this.redraw()),"auto"==t&&(t=e.call(this))),this.nextStyle=t},t.fallback=function(t){var n=this.el.style[this.name]||this.convert(this.get(),this.type);t=this.convert(t,this.type),this.auto&&("auto"==n&&(n=this.convert(this.get(),this.type)),"auto"==t&&(t=e.call(this))),this.tween=new W({from:n,to:t,duration:this.duration,delay:this.delay,ease:this.ease,update:this.update,context:this})},t.get=function(){return X(this.el,this.name)},t.update=function(t){G(this.el,this.name,t)},t.stop=function(){(this.active||this.nextStyle)&&(this.active=!1,this.nextStyle=null,G(this.el,this.name,this.get()));var t=this.tween;t&&t.context&&t.destroy()},t.convert=function(t,e){if("auto"==t&&this.auto)return t;var r,o="number"==typeof t,a="string"==typeof t;switch(e){case g:if(o)return t;if(a&&""===t.replace(m,""))return+t;r="number(unitless)";break;case b:if(a){if(""===t&&this.original)return this.original;if(e.test(t))return"#"==t.charAt(0)&&7==t.length?t:n(t)}r="hex or rgb string";break;case y:if(o)return t+this.unit;if(a&&e.test(t))return t;r="number(px) or string(unit)";break;case x:if(o)return t+this.unit;if(a&&e.test(t))return t;r="number(px) or string(unit or %)";break;case k:if(o)return t+this.angle;if(a&&e.test(t))return t;r="number(deg) or string(angle)";break;case _:if(o)return t;if(a&&x.test(t))return t;r="number(unitless) or string(unit or %)"}return function(t,e){c("Type warning: Expected: ["+t+"] Got: ["+(0,i.default)(e)+"] "+e)}(r,t),t},t.redraw=function(){this.el.offsetHeight}}),F=l(N,function(t,e){t.init=function(){e.init.apply(this,arguments),this.original||(this.original=this.convert(this.get(),b))}}),$=l(N,function(t,e){t.init=function(){e.init.apply(this,arguments),this.animate=this.fallback},t.get=function(){return this.$el[this.name]()},t.update=function(t){this.$el[this.name](t)}}),q=l(N,function(t,e){function n(t,e){var n,i,r,o,a;for(n in t)r=(o=K[n])[0],i=o[1]||n,a=this.convert(t[n],r),e.call(this,i,a,r)}t.init=function(){e.init.apply(this,arguments),this.current||(this.current={},K.perspective&&B.perspective&&(this.current.perspective=B.perspective,G(this.el,this.name,this.style(this.current)),this.redraw()))},t.set=function(t){n.call(this,t,function(t,e){this.current[t]=e}),G(this.el,this.name,this.style(this.current)),this.redraw()},t.transition=function(t){var e=this.values(t);this.tween=new U({current:this.current,values:e,duration:this.duration,delay:this.delay,ease:this.ease});var n,i={};for(n in this.current)i[n]=n in e?e[n]:this.current[n];this.active=!0,this.nextStyle=this.style(i)},t.fallback=function(t){var e=this.values(t);this.tween=new U({current:this.current,values:e,duration:this.duration,delay:this.delay,ease:this.ease,update:this.update,context:this})},t.update=function(){G(this.el,this.name,this.style(this.current))},t.style=function(t){var e,n="";for(e in t)n+=e+"("+t[e]+") ";return n},t.values=function(t){var e,i={};return n.call(this,t,function(t,n,r){i[t]=n,void 0===this.current[t]&&(e=0,~t.indexOf("scale")&&(e=1),this.current[t]=this.convert(e,r))}),i}}),W=l(function(e){function n(){var t,e,i,r=u.length;if(r)for(I(n),e=P(),t=r;t--;)(i=u[t])&&i.render(e)}var i={ease:f.ease[1],from:0,to:1};e.init=function(t){this.duration=t.duration||0,this.delay=t.delay||0;var e=t.ease||i.ease;f[e]&&(e=f[e][1]),"function"!=typeof e&&(e=i.ease),this.ease=e,this.update=t.update||a,this.complete=t.complete||a,this.context=t.context||this,this.name=t.name;var n=t.from,r=t.to;void 0===n&&(n=i.from),void 0===r&&(r=i.to),this.unit=t.unit||"","number"==typeof n&&"number"==typeof r?(this.begin=n,this.change=r-n):this.format(r,n),this.value=this.begin+this.unit,this.start=P(),!1!==t.autoplay&&this.play()},e.play=function(){var t;this.active||(this.start||(this.start=P()),this.active=!0,t=this,1===u.push(t)&&I(n))},e.stop=function(){var e,n,i;this.active&&(this.active=!1,e=this,(i=t.inArray(e,u))>=0&&(n=u.slice(i+1),u.length=i,n.length&&(u=u.concat(n))))},e.render=function(t){var e,n=t-this.start;if(this.delay){if(n<=this.delay)return;n-=this.delay}if(n<this.duration){var i=this.ease(n,0,1,this.duration);return e=this.startRGB?function(t,e,n){return o(t[0]+n*(e[0]-t[0]),t[1]+n*(e[1]-t[1]),t[2]+n*(e[2]-t[2]))}(this.startRGB,this.endRGB,i):function(t){return Math.round(t*c)/c}(this.begin+i*this.change),this.value=e+this.unit,void this.update.call(this.context,this.value)}e=this.endHex||this.begin+this.change,this.value=e+this.unit,this.update.call(this.context,this.value),this.complete.call(this.context),this.destroy()},e.format=function(t,e){if(e+="","#"==(t+="").charAt(0))return this.startRGB=r(e),this.endRGB=r(t),this.endHex=t,this.begin=0,void(this.change=1);if(!this.unit){var n=e.replace(m,"");n!==t.replace(m,"")&&s("tween",e,t),this.unit=n}e=parseFloat(e),t=parseFloat(t),this.begin=this.value=e,this.change=t-e},e.destroy=function(){this.stop(),this.context=null,this.ease=this.update=this.complete=a};var u=[],c=1e3}),H=l(W,function(t){t.init=function(t){this.duration=t.duration||0,this.complete=t.complete||a,this.context=t.context,this.play()},t.render=function(t){t-this.start<this.duration||(this.complete.call(this.context),this.destroy())}}),U=l(W,function(t,e){t.init=function(t){var e,n;for(e in this.context=t.context,this.update=t.update,this.tweens=[],this.current=t.current,t.values)n=t.values[e],this.current[e]!==n&&this.tweens.push(new W({name:e,from:this.current[e],to:n,duration:t.duration,delay:t.delay,ease:t.ease,autoplay:!1}));this.play()},t.render=function(t){var e,n,i=!1;for(e=this.tweens.length;e--;)(n=this.tweens[e]).context&&(n.render(t),this.current[n.name]=n.value,i=!0);return i?void(this.update&&this.update.call(this.context)):this.destroy()},t.destroy=function(){if(e.destroy.call(this),this.tweens){var t;for(t=this.tweens.length;t--;)this.tweens[t].destroy();this.tweens=null,this.current=null}}}),B=e.config={debug:!1,defaultUnit:"px",defaultAngle:"deg",keepInherited:!1,hideBackface:!1,perspective:"",fallback:!z.transition,agentTests:[]};e.fallback=function(t){if(!z.transition)return B.fallback=!0;B.agentTests.push("("+t+")");var e=new RegExp(B.agentTests.join("|"),"i");B.fallback=e.test(navigator.userAgent)},e.fallback("6.0.[2-5] Safari"),e.tween=function(t){return new W(t)},e.delay=function(t,e,n){return new H({complete:e,duration:t,context:n})},t.fn.tram=function(t){return e.call(null,this,t)};var G=t.style,X=t.css,Z={transform:z.transform&&z.transform.css},Y={color:[F,b],background:[F,b,"background-color"],"outline-color":[F,b],"border-color":[F,b],"border-top-color":[F,b],"border-right-color":[F,b],"border-bottom-color":[F,b],"border-left-color":[F,b],"border-width":[N,y],"border-top-width":[N,y],"border-right-width":[N,y],"border-bottom-width":[N,y],"border-left-width":[N,y],"border-spacing":[N,y],"letter-spacing":[N,y],margin:[N,y],"margin-top":[N,y],"margin-right":[N,y],"margin-bottom":[N,y],"margin-left":[N,y],padding:[N,y],"padding-top":[N,y],"padding-right":[N,y],"padding-bottom":[N,y],"padding-left":[N,y],"outline-width":[N,y],opacity:[N,g],top:[N,x],right:[N,x],bottom:[N,x],left:[N,x],"font-size":[N,x],"text-indent":[N,x],"word-spacing":[N,x],width:[N,x],"min-width":[N,x],"max-width":[N,x],height:[N,x],"min-height":[N,x],"max-height":[N,x],"line-height":[N,_],"scroll-top":[$,g,"scrollTop"],"scroll-left":[$,g,"scrollLeft"]},K={};z.transform&&(Y.transform=[q],K={x:[x,"translateX"],y:[x,"translateY"],rotate:[k],rotateX:[k],rotateY:[k],scale:[g],scaleX:[g],scaleY:[g],skew:[k],skewX:[k],skewY:[k]}),z.transform&&z.backface&&(K.z=[x,"translateZ"],K.rotateZ=[k],K.scaleZ=[g],K.perspective=[y]);var Q=/ms/,J=/s|\./;return t.tram=e}(window.jQuery)},function(t,e){t.exports=function(t){return t&&t.__esModule?t:{default:t}}},function(t,e,n){n(4),n(7),n(8),n(9),n(10),t.exports=n(15)},function(t,e,n){"use strict";var i=n(0);i.define("brand",t.exports=function(t){var e,n={},r=document,o=t("html"),a=t("body"),s=".w-webflow-badge",u=window.location,c=/PhantomJS/i.test(navigator.userAgent),l="fullscreenchange webkitfullscreenchange mozfullscreenchange msfullscreenchange";function f(){var n=r.fullScreen||r.mozFullScreen||r.webkitIsFullScreen||r.msFullscreenElement||Boolean(r.webkitFullscreenElement);t(e).attr("style",n?"display: none !important;":"")}function d(){var t=a.children(s),n=t.length&&t.get(0)===e,r=i.env("editor");n?r&&t.remove():(t.length&&t.remove(),r||a.append(e))}return n.ready=function(){var n,i,a,s=o.attr("data-wf-status"),h=o.attr("data-wf-domain")||"";/\.webflow\.io$/i.test(h)&&u.hostname!==h&&(s=!0),s&&!c&&(e=e||(n=t('<a class="w-webflow-badge"></a>').attr("href","https://webflow.com?utm_campaign=brandjs"),i=t("<img>").attr("src","https://d3e54v103j8qbb.cloudfront.net/img/webflow-badge-icon.f67cd735e3.svg").attr("alt","").css({marginRight:"8px",width:"16px"}),a=t("<img>").attr("src","https://d1otoma47x30pg.cloudfront.net/img/webflow-badge-text.6faa6a38cd.svg").attr("alt","Made in Webflow"),n.append(i,a),n[0]),d(),setTimeout(d,500),t(r).off(l,f).on(l,f))},n})},function(t,e,n){"use strict";var i=window.$,r=n(1)&&i.tram;
-/*!
- * Webflow._ (aka) Underscore.js 1.6.0 (custom build)
- * _.each
- * _.map
- * _.find
- * _.filter
- * _.any
- * _.contains
- * _.delay
- * _.defer
- * _.throttle (webflow)
- * _.debounce
- * _.keys
- * _.has
- * _.now
- *
- * http://underscorejs.org
- * (c) 2009-2013 Jeremy Ashkenas, DocumentCloud and Investigative Reporters & Editors
- * Underscore may be freely distributed under the MIT license.
- * @license MIT
- */
-t.exports=function(){var t={VERSION:"1.6.0-Webflow"},e={},n=Array.prototype,i=Object.prototype,o=Function.prototype,a=(n.push,n.slice),s=(n.concat,i.toString,i.hasOwnProperty),u=n.forEach,c=n.map,l=(n.reduce,n.reduceRight,n.filter),f=(n.every,n.some),d=n.indexOf,h=(n.lastIndexOf,Array.isArray,Object.keys),p=(o.bind,t.each=t.forEach=function(n,i,r){if(null==n)return n;if(u&&n.forEach===u)n.forEach(i,r);else if(n.length===+n.length){for(var o=0,a=n.length;o<a;o++)if(i.call(r,n[o],o,n)===e)return}else{var s=t.keys(n);for(o=0,a=s.length;o<a;o++)if(i.call(r,n[s[o]],s[o],n)===e)return}return n});t.map=t.collect=function(t,e,n){var i=[];return null==t?i:c&&t.map===c?t.map(e,n):(p(t,function(t,r,o){i.push(e.call(n,t,r,o))}),i)},t.find=t.detect=function(t,e,n){var i;return v(t,function(t,r,o){if(e.call(n,t,r,o))return i=t,!0}),i},t.filter=t.select=function(t,e,n){var i=[];return null==t?i:l&&t.filter===l?t.filter(e,n):(p(t,function(t,r,o){e.call(n,t,r,o)&&i.push(t)}),i)};var v=t.some=t.any=function(n,i,r){i||(i=t.identity);var o=!1;return null==n?o:f&&n.some===f?n.some(i,r):(p(n,function(t,n,a){if(o||(o=i.call(r,t,n,a)))return e}),!!o)};t.contains=t.include=function(t,e){return null!=t&&(d&&t.indexOf===d?-1!=t.indexOf(e):v(t,function(t){return t===e}))},t.delay=function(t,e){var n=a.call(arguments,2);return setTimeout(function(){return t.apply(null,n)},e)},t.defer=function(e){return t.delay.apply(t,[e,1].concat(a.call(arguments,1)))},t.throttle=function(t){var e,n,i;return function(){e||(e=!0,n=arguments,i=this,r.frame(function(){e=!1,t.apply(i,n)}))}},t.debounce=function(e,n,i){var r,o,a,s,u,c=function c(){var l=t.now()-s;l<n?r=setTimeout(c,n-l):(r=null,i||(u=e.apply(a,o),a=o=null))};return function(){a=this,o=arguments,s=t.now();var l=i&&!r;return r||(r=setTimeout(c,n)),l&&(u=e.apply(a,o),a=o=null),u}},t.defaults=function(e){if(!t.isObject(e))return e;for(var n=1,i=arguments.length;n<i;n++){var r=arguments[n];for(var o in r)void 0===e[o]&&(e[o]=r[o])}return e},t.keys=function(e){if(!t.isObject(e))return[];if(h)return h(e);var n=[];for(var i in e)t.has(e,i)&&n.push(i);return n},t.has=function(t,e){return s.call(t,e)},t.isObject=function(t){return t===Object(t)},t.now=Date.now||function(){return(new Date).getTime()},t.templateSettings={evaluate:/<%([\s\S]+?)%>/g,interpolate:/<%=([\s\S]+?)%>/g,escape:/<%-([\s\S]+?)%>/g};var m=/(.)^/,w={"'":"'","\\":"\\","\r":"r","\n":"n","\u2028":"u2028","\u2029":"u2029"},g=/\\|'|\r|\n|\u2028|\u2029/g,b=function(t){return"\\"+w[t]};return t.template=function(e,n,i){!n&&i&&(n=i),n=t.defaults({},n,t.templateSettings);var r=RegExp([(n.escape||m).source,(n.interpolate||m).source,(n.evaluate||m).source].join("|")+"|$","g"),o=0,a="__p+='";e.replace(r,function(t,n,i,r,s){return a+=e.slice(o,s).replace(g,b),o=s+t.length,n?a+="'+\n((__t=("+n+"))==null?'':_.escape(__t))+\n'":i?a+="'+\n((__t=("+i+"))==null?'':__t)+\n'":r&&(a+="';\n"+r+"\n__p+='"),t}),a+="';\n",n.variable||(a="with(obj||{}){\n"+a+"}\n"),a="var __t,__p='',__j=Array.prototype.join,print=function(){__p+=__j.call(arguments,'');};\n"+a+"return __p;\n";try{var s=new Function(n.variable||"obj","_",a)}catch(t){throw t.source=a,t}var u=function(e){return s.call(this,e,t)},c=n.variable||"obj";return u.source="function("+c+"){\n"+a+"}",u},t}()},function(t,e){function n(t){return(n="function"==typeof Symbol&&"symbol"==typeof Symbol.iterator?function(t){return typeof t}:function(t){return t&&"function"==typeof Symbol&&t.constructor===Symbol&&t!==Symbol.prototype?"symbol":typeof t})(t)}function i(e){return"function"==typeof Symbol&&"symbol"===n(Symbol.iterator)?t.exports=i=function(t){return n(t)}:t.exports=i=function(t){return t&&"function"==typeof Symbol&&t.constructor===Symbol&&t!==Symbol.prototype?"symbol":n(t)},i(e)}t.exports=i},function(t,e,n){"use strict";var i=n(0);i.define("links",t.exports=function(t,e){var n,r,o,a={},s=t(window),u=i.env(),c=window.location,l=document.createElement("a"),f="w--current",d=/index\.(html|php)$/,h=/\/$/;function p(e){var i=n&&e.getAttribute("href-disabled")||e.getAttribute("href");if(l.href=i,!(i.indexOf(":")>=0)){var a=t(e);if(l.hash.length>1&&l.host+l.pathname===c.host+c.pathname){if(!/^#[a-zA-Z0-9\-\_]+$/.test(l.hash))return;var s=t(l.hash);s.length&&r.push({link:a,sec:s,active:!1})}else if("#"!==i&&""!==i){var u=l.href===c.href||i===o||d.test(i)&&h.test(o);m(a,f,u)}}}function v(){var t=s.scrollTop(),n=s.height();e.each(r,function(e){var i=e.link,r=e.sec,o=r.offset().top,a=r.outerHeight(),s=.5*n,u=r.is(":visible")&&o+a-s>=t&&o+s<=t+n;e.active!==u&&(e.active=u,m(i,f,u))})}function m(t,e,n){var i=t.hasClass(e);n&&i||(n||i)&&(n?t.addClass(e):t.removeClass(e))}return a.ready=a.design=a.preview=function(){n=u&&i.env("design"),o=i.env("slug")||c.pathname||"",i.scroll.off(v),r=[];for(var t=document.links,e=0;e<t.length;++e)p(t[e]);r.length&&(i.scroll.on(v),v())},a})},function(t,e,n){"use strict";var i=n(0);i.define("scroll",t.exports=function(t){var e={WF_CLICK_EMPTY:"click.wf-empty-link",WF_CLICK_SCROLL:"click.wf-scroll"},n=window.location,r=function(){try{return Boolean(window.frameElement)}catch(t){return!0}}()?null:window.history,o=t(window),a=t(document),s=t(document.body),u=window.requestAnimationFrame||window.mozRequestAnimationFrame||window.webkitRequestAnimationFrame||function(t){window.setTimeout(t,15)},c=i.env("editor")?".w-editor-body":"body",l="header, "+c+" > .header, "+c+" > .w-nav:not([data-no-scroll])",f='a[href="#"]',d='a[href*="#"]:not(.w-tab-link):not('+f+")";var h=/^#[a-zA-Z0-9][\w:.-]*$/;function p(e){var a=e.currentTarget;if(!(i.env("design")||window.$.mobile&&/(?:^|\s)ui-link(?:$|\s)/.test(a.className))){var c,f=(c=a,h.test(c.hash)&&c.host+c.pathname===n.host+n.pathname?a.hash:"");if(""!==f){var d=t(f);d.length&&(e&&(e.preventDefault(),e.stopPropagation()),function(t){if(n.hash!==t&&r&&r.pushState&&(!i.env.chrome||"file:"!==n.protocol)){var e=r.state&&r.state.hash;e!==t&&r.pushState({hash:t},"",t)}}(f),window.setTimeout(function(){!function(e){var n=o.scrollTop(),i=function(e){var n=t(l),i="fixed"===n.css("position")?n.outerHeight():0,r=e.offset().top-i;if("mid"===e.data("scroll")){var a=o.height()-i,s=e.outerHeight();s<a&&(r-=Math.round((a-s)/2))}return r}(e);if(n===i)return;var r=function(t,e,n){if(document.body.hasAttribute("data-wf-reduce-scroll-motion")&&("none"===document.body.getAttribute("data-wf-scroll-motion")||"function"==typeof window.matchMedia&&window.matchMedia("(prefers-reduced-motion: reduce)").matches))return 0;var i=1;return s.add(t).each(function(t,e){var n=parseFloat(e.getAttribute("data-scroll-time"));!isNaN(n)&&n>=0&&(i=n)}),(472.143*Math.log(Math.abs(e-n)+125)-2e3)*i}(e,n,i),a=Date.now();u(function t(){var e=Date.now()-a;window.scroll(0,function(t,e,n,i){return n>i?e:t+(e-t)*((r=n/i)<.5?4*r*r*r:(r-1)*(2*r-2)*(2*r-2)+1);var r}(n,i,e,r)),e<=r&&u(t)})}(d)},e?0:300))}}}return{ready:function(){var t=e.WF_CLICK_EMPTY,n=e.WF_CLICK_SCROLL;a.on(n,d,p),a.on(t,f,function(t){t.preventDefault()})}}})},function(t,e,n){"use strict";n(0).define("touch",t.exports=function(t){var e={},n=window.getSelection;function i(e){var i,r,o=!1,a=!1,s=Math.min(Math.round(.04*window.innerWidth),40);function u(t){var e=t.touches;e&&e.length>1||(o=!0,e?(a=!0,i=e[0].clientX):i=t.clientX,r=i)}function c(e){if(o){if(a&&"mousemove"===e.type)return e.preventDefault(),void e.stopPropagation();var i=e.touches,u=i?i[0].clientX:e.clientX,c=u-r;r=u,Math.abs(c)>s&&n&&""===String(n())&&(!function(e,n,i){var r=t.Event(e,{originalEvent:n});t(n.target).trigger(r,i)}("swipe",e,{direction:c>0?"right":"left"}),f())}}function l(t){if(o)return o=!1,a&&"mouseup"===t.type?(t.preventDefault(),t.stopPropagation(),void(a=!1)):void 0}function f(){o=!1}e.addEventListener("touchstart",u,!1),e.addEventListener("touchmove",c,!1),e.addEventListener("touchend",l,!1),e.addEventListener("touchcancel",f,!1),e.addEventListener("mousedown",u,!1),e.addEventListener("mousemove",c,!1),e.addEventListener("mouseup",l,!1),e.addEventListener("mouseout",f,!1),this.destroy=function(){e.removeEventListener("touchstart",u,!1),e.removeEventListener("touchmove",c,!1),e.removeEventListener("touchend",l,!1),e.removeEventListener("touchcancel",f,!1),e.removeEventListener("mousedown",u,!1),e.removeEventListener("mousemove",c,!1),e.removeEventListener("mouseup",l,!1),e.removeEventListener("mouseout",f,!1),e=null}}return t.event.special.tap={bindType:"click",delegateType:"click"},e.init=function(e){return(e="string"==typeof e?t(e).get(0):e)?new i(e):null},e.instance=e.init(document),e})},function(t,e,n){"use strict";var i=n(2)(n(11)),r=n(0);r.define("forms",t.exports=function(t,e){var n,o,a,s,u,c={},l=t(document),f=window.location,d=window.XDomainRequest&&!window.atob,h=".w-form",p=/e(-)?mail/i,v=/^\S+@\S+$/,m=window.alert,w=r.env(),g=/list-manage[1-9]?.com/i,b=e.debounce(function(){m("Oops! This page has improperly configured forms. Please contact your website administrator to fix this issue.")},100);function y(e,n){var i=t(n),r=t.data(n,h);r||(r=t.data(n,h,{form:i})),x(r);var a=i.closest("div.w-form");r.done=a.find("> .w-form-done"),r.fail=a.find("> .w-form-fail"),r.fileUploads=a.find(".w-file-upload"),r.fileUploads.each(function(e){!function(e,n){if(!n.fileUploads||!n.fileUploads[e])return;var i,r=t(n.fileUploads[e]),o=r.find("> .w-file-upload-default"),a=r.find("> .w-file-upload-uploading"),s=r.find("> .w-file-upload-success"),c=r.find("> .w-file-upload-error"),l=o.find(".w-file-upload-input"),f=o.find(".w-file-upload-label"),d=f.children(),h=c.find(".w-file-upload-error-msg"),p=s.find(".w-file-upload-file"),v=s.find(".w-file-remove-link"),m=p.find(".w-file-upload-file-name"),g=h.attr("data-w-size-error"),b=h.attr("data-w-type-error"),y=h.attr("data-w-generic-error");if(w)l.on("click",function(t){t.preventDefault()}),f.on("click",function(t){t.preventDefault()}),d.on("click",function(t){t.preventDefault()});else{v.on("click",function(){l.removeAttr("data-value"),l.val(""),m.html(""),o.toggle(!0),s.toggle(!1)}),l.on("change",function(r){(i=r.target&&r.target.files&&r.target.files[0])&&(o.toggle(!1),c.toggle(!1),a.toggle(!0),m.text(i.name),S()||k(n),n.fileUploads[e].uploading=!0,function(e,n){var i={name:e.name,size:e.size};t.ajax({type:"POST",url:u,data:i,dataType:"json",crossDomain:!0}).done(function(t){n(null,t)}).fail(function(t){n(t)})}(i,O))});var _=f.outerHeight();l.height(_),l.width(1)}function E(t){var i=t.responseJSON&&t.responseJSON.msg,r=y;"string"==typeof i&&0===i.indexOf("InvalidFileTypeError")?r=b:"string"==typeof i&&0===i.indexOf("MaxFileSizeError")&&(r=g),h.text(r),l.removeAttr("data-value"),l.val(""),a.toggle(!1),o.toggle(!0),c.toggle(!0),n.fileUploads[e].uploading=!1,S()||x(n)}function O(e,n){if(e)return E(e);var r=n.fileName,o=n.postData,a=n.fileId,s=n.s3Url;l.attr("data-value",a),function(e,n,i,r,o){var a=new FormData;for(var s in n)a.append(s,n[s]);a.append("file",i,r),t.ajax({type:"POST",url:e,data:a,processData:!1,contentType:!1}).done(function(){o(null)}).fail(function(t){o(t)})}(s,o,i,r,A)}function A(t){if(t)return E(t);a.toggle(!1),s.css("display","inline-block"),n.fileUploads[e].uploading=!1,S()||x(n)}function S(){var t=n.fileUploads&&n.fileUploads.toArray()||[];return t.some(function(t){return t.uploading})}}(e,r)});var s=r.action=i.attr("action");r.handler=null,r.redirect=i.attr("data-redirect"),g.test(s)?r.handler=O:s||(o?r.handler="function"==typeof hostedSubmitWebflow?hostedSubmitWebflow:E:b())}function x(t){var e=t.btn=t.form.find(':input[type="submit"]');t.wait=t.btn.attr("data-wait")||null,t.success=!1,e.prop("disabled",!1),t.label&&e.val(t.label)}function k(t){var e=t.btn,n=t.wait;e.prop("disabled",!0),n&&(t.label=e.val(),e.val(n))}function _(e,n){var i=null;return n=n||{},e.find(':input:not([type="submit"]):not([type="file"])').each(function(r,o){var a=t(o),s=a.attr("type"),u=a.attr("data-name")||a.attr("name")||"Field "+(r+1),c=a.val();if("checkbox"===s)c=a.is(":checked");else if("radio"===s){if(null===n[u]||"string"==typeof n[u])return;c=e.find('input[name="'+a.attr("name")+'"]:checked').val()||null}"string"==typeof c&&(c=t.trim(c)),n[u]=c,i=i||function(t,e,n,i){var r=null;"password"===e?r="Passwords cannot be submitted.":t.attr("required")?i?p.test(t.attr("type"))&&(v.test(i)||(r="Please enter a valid email address for: "+n)):r="Please fill out the required field: "+n:"g-recaptcha-response"!==n||i||(r="Please confirm you’re not a robot.");return r}(a,s,u,c)}),i}function E(t){S(t),A(t)}function O(n){x(n);var i=n.form,r={};if(!/^https/.test(f.href)||/^https/.test(n.action)){S(n);var o,a=_(i,r);if(a)return m(a);k(n),e.each(r,function(t,e){p.test(e)&&(r.EMAIL=t),/^((full[ _-]?)?name)$/i.test(e)&&(o=t),/^(first[ _-]?name)$/i.test(e)&&(r.FNAME=t),/^(last[ _-]?name)$/i.test(e)&&(r.LNAME=t)}),o&&!r.FNAME&&(o=o.split(" "),r.FNAME=o[0],r.LNAME=r.LNAME||o[1]);var s=n.action.replace("/post?","/post-json?")+"&c=?",u=s.indexOf("u=")+2;u=s.substring(u,s.indexOf("&",u));var c=s.indexOf("id=")+3;c=s.substring(c,s.indexOf("&",c)),r["b_"+u+"_"+c]="",t.ajax({url:s,data:r,dataType:"jsonp"}).done(function(t){n.success="success"===t.result||/already/.test(t.msg),n.success||console.info("MailChimp error: "+t.msg),A(n)}).fail(function(){A(n)})}else i.attr("method","post")}function A(t){var e=t.form,n=t.redirect,i=t.success;i&&n?r.location(n):(t.done.toggle(i),t.fail.toggle(!i),e.toggle(!i),x(t))}function S(t){t.evt&&t.evt.preventDefault(),t.evt=null}return c.ready=c.design=c.preview=function(){!function(){o=t("html").attr("data-wf-site"),s="https://webflow.com/api/v1/form/"+o,d&&s.indexOf("https://webflow.com")>=0&&(s=s.replace("https://webflow.com","http://formdata.webflow.com"));if(u="".concat(s,"/signFile"),!(n=t(h+" form")).length)return;n.each(y)}(),w||a||function(){a=!0,l.on("submit",h+" form",function(e){var n=t.data(this,h);n.handler&&(n.evt=e,n.handler(n))});var e=[["checkbox",".w-checkbox-input"],["radio",".w-radio-input"]];l.on("change",h+' form input[type="checkbox"]:not(.w-checkbox-input)',function(e){t(e.target).siblings(".w-checkbox-input").toggleClass("w--redirected-checked")}),l.on("change",h+' form input[type="radio"]',function(e){t('input[name="'.concat(e.target.name,'"]:not(').concat(".w-checkbox-input",")")).map(function(e,n){return t(n).siblings(".w-radio-input").removeClass("w--redirected-checked")});var n=t(e.target);n.hasClass("w-radio-input")||n.siblings(".w-radio-input").addClass("w--redirected-checked")}),e.forEach(function(e){var n=(0,i.default)(e,2),r=n[0],o=n[1];l.on("focus",h+' form input[type="'.concat(r,'"]:not(')+o+")",function(e){t(e.target).siblings(o).addClass("w--redirected-focus")}),l.on("blur",h+' form input[type="'.concat(r,'"]:not(')+o+")",function(e){t(e.target).siblings(o).removeClass("w--redirected-focus")})})}()},c})},function(t,e,n){var i=n(12),r=n(13),o=n(14);t.exports=function(t,e){return i(t)||r(t,e)||o()}},function(t,e){t.exports=function(t){if(Array.isArray(t))return t}},function(t,e){t.exports=function(t,e){var n=[],i=!0,r=!1,o=void 0;try{for(var a,s=t[Symbol.iterator]();!(i=(a=s.next()).done)&&(n.push(a.value),!e||n.length!==e);i=!0);}catch(t){r=!0,o=t}finally{try{i||null==s.return||s.return()}finally{if(r)throw o}}return n}},function(t,e){t.exports=function(){throw new TypeError("Invalid attempt to destructure non-iterable instance")}},function(t,e,n){"use strict";var i=n(0),r=n(16),o={ARROW_LEFT:37,ARROW_UP:38,ARROW_RIGHT:39,ARROW_DOWN:40,ESCAPE:27,SPACE:32,ENTER:13,HOME:36,END:35};i.define("navbar",t.exports=function(t,e){var n,a,s,u,c={},l=t.tram,f=t(window),d=t(document),h=e.debounce,p=i.env(),v='<div class="w-nav-overlay" data-wf-ignore />',m=".w-nav",w="w--open",g="w--nav-dropdown-open",b="w--nav-dropdown-toggle-open",y="w--nav-dropdown-list-open",x="w--nav-link-open",k=r.triggers,_=t();function E(){i.resize.off(O)}function O(){a.each(P)}function A(n,i){var r=t(i),a=t.data(i,m);a||(a=t.data(i,m,{open:!1,el:r,config:{},selectedIdx:-1})),a.menu=r.find(".w-nav-menu"),a.links=a.menu.find(".w-nav-link"),a.dropdowns=a.menu.find(".w-dropdown"),a.dropdownToggle=a.menu.find(".w-dropdown-toggle"),a.dropdownList=a.menu.find(".w-dropdown-list"),a.button=r.find(".w-nav-button"),a.container=r.find(".w-container"),a.overlayContainerId="w-nav-overlay-"+n,a.outside=function(e){e.outside&&d.off("click"+m,e.outside);return function(n){var i=t(n.target);u&&i.closest(".w-editor-bem-EditorOverlay").length||I(e,i)}}(a);var c=r.find(".w-nav-brand");c&&"/"===c.attr("href")&&null==c.attr("aria-label")&&c.attr("aria-label","home"),a.button.attr("style","-webkit-user-select: text;"),null==a.button.attr("aria-label")&&a.button.attr("aria-label","menu"),a.button.attr("role","button"),a.button.attr("tabindex","0"),a.button.attr("aria-controls",a.overlayContainerId),a.button.attr("aria-haspopup","menu"),a.button.attr("aria-expanded","false"),a.el.off(m),a.button.off(m),a.menu.off(m),C(a),s?(T(a),a.el.on("setting"+m,function(t){return function(n,i){i=i||{};var r=f.width();C(t),!0===i.open&&F(t,!0),!1===i.open&&q(t,!0),t.open&&e.defer(function(){r!==f.width()&&z(t)})}}(a))):(!function(e){if(e.overlay)return;e.overlay=t(v).appendTo(e.el),e.overlay.attr("id",e.overlayContainerId),e.parent=e.menu.parent(),q(e,!0)}(a),a.button.on("click"+m,M(a)),a.menu.on("click"+m,"a",D(a)),a.button.on("keydown"+m,function(t){return function(e){switch(e.keyCode){case o.SPACE:case o.ENTER:return M(t)(),e.preventDefault(),e.stopPropagation();case o.ESCAPE:return q(t),e.preventDefault(),e.stopPropagation();case o.ARROW_RIGHT:case o.ARROW_DOWN:case o.HOME:case o.END:return t.open?(e.keyCode===o.END?t.selectedIdx=t.links.length-1:t.selectedIdx=0,R(t),e.preventDefault(),e.stopPropagation()):(e.preventDefault(),e.stopPropagation())}}}(a)),a.el.on("keydown"+m,function(t){return function(e){if(t.open)switch(t.selectedIdx=t.links.index(document.activeElement),e.keyCode){case o.HOME:case o.END:return e.keyCode===o.END?t.selectedIdx=t.links.length-1:t.selectedIdx=0,R(t),e.preventDefault(),e.stopPropagation();case o.ESCAPE:return q(t),t.button.focus(),e.preventDefault(),e.stopPropagation();case o.ARROW_LEFT:case o.ARROW_UP:return t.selectedIdx=Math.max(-1,t.selectedIdx-1),R(t),e.preventDefault(),e.stopPropagation();case o.ARROW_RIGHT:case o.ARROW_DOWN:return t.selectedIdx=Math.min(t.links.length-1,t.selectedIdx+1),R(t),e.preventDefault(),e.stopPropagation()}}}(a))),P(n,i)}function S(e,n){var i=t.data(n,m);i&&(T(i),t.removeData(n,m))}function T(t){t.overlay&&(q(t,!0),t.overlay.remove(),t.overlay=null)}function C(t){var n={},i=t.config||{},r=n.animation=t.el.attr("data-animation")||"default";n.animOver=/^over/.test(r),n.animDirect=/left$/.test(r)?-1:1,i.animation!==r&&t.open&&e.defer(z,t),n.easing=t.el.attr("data-easing")||"ease",n.easing2=t.el.attr("data-easing2")||"ease";var o=t.el.attr("data-duration");n.duration=null!=o?Number(o):400,n.docHeight=t.el.attr("data-doc-height"),t.config=n}function R(t){if(t.links[t.selectedIdx]){var e=t.links[t.selectedIdx];e.focus(),D(e)}}function z(t){t.open&&(q(t,!0),F(t,!0))}function M(t){return h(function(){t.open?q(t):F(t)})}function D(e){return function(n){var r=t(this).attr("href");i.validClick(n.currentTarget)?r&&0===r.indexOf("#")&&e.open&&q(e):n.preventDefault()}}c.ready=c.design=c.preview=function(){if(s=p&&i.env("design"),u=i.env("editor"),n=t(document.body),!(a=d.find(m)).length)return;a.each(A),E(),i.resize.on(O)},c.destroy=function(){_=t(),E(),a&&a.length&&a.each(S)};var I=h(function(t,e){if(t.open){var n=e.closest(".w-nav-menu");t.menu.is(n)||q(t)}});function P(e,n){var i=t.data(n,m),r=i.collapsed="none"!==i.button.css("display");if(!i.open||r||s||q(i,!0),i.container.length){var o=function(e){var n=e.container.css(j);"none"===n&&(n="");return function(e,i){(i=t(i)).css(j,""),"none"===i.css(j)&&i.css(j,n)}}(i);i.links.each(o),i.dropdowns.each(o)}i.open&&$(i)}var j="max-width";function L(t,e){e.setAttribute("data-nav-menu-open","")}function N(t,e){e.removeAttribute("data-nav-menu-open")}function F(t,e){if(!t.open){t.open=!0,t.menu.each(L),t.links.addClass(x),t.dropdowns.addClass(g),t.dropdownToggle.addClass(b),t.dropdownList.addClass(y),t.button.addClass(w);var n=t.config;("none"===n.animation||!l.support.transform||n.duration<=0)&&(e=!0);var r=$(t),o=t.menu.outerHeight(!0),a=t.menu.outerWidth(!0),u=t.el.height(),c=t.el[0];if(P(0,c),k.intro(0,c),i.redraw.up(),s||d.on("click"+m,t.outside),e)p();else{var f="transform "+n.duration+"ms "+n.easing;if(t.overlay&&(_=t.menu.prev(),t.overlay.show().append(t.menu)),n.animOver)return l(t.menu).add(f).set({x:n.animDirect*a,height:r}).start({x:0}).then(p),void(t.overlay&&t.overlay.width(a));var h=u+o;l(t.menu).add(f).set({y:-h}).start({y:0}).then(p)}}function p(){t.button.attr("aria-expanded","true")}}function $(t){var e=t.config,i=e.docHeight?d.height():n.height();return e.animOver?t.menu.height(i):"fixed"!==t.el.css("position")&&(i-=t.el.outerHeight(!0)),t.overlay&&t.overlay.height(i),i}function q(t,e){if(t.open){t.open=!1,t.button.removeClass(w);var n=t.config;if(("none"===n.animation||!l.support.transform||n.duration<=0)&&(e=!0),k.outro(0,t.el[0]),d.off("click"+m,t.outside),e)return l(t.menu).stop(),void u();var i="transform "+n.duration+"ms "+n.easing2,r=t.menu.outerHeight(!0),o=t.menu.outerWidth(!0),a=t.el.height();if(n.animOver)l(t.menu).add(i).start({x:o*n.animDirect}).then(u);else{var s=a+r;l(t.menu).add(i).start({y:-s}).then(u)}}function u(){t.menu.height(""),l(t.menu).set({x:0,y:0}),t.menu.each(N),t.links.removeClass(x),t.dropdowns.removeClass(g),t.dropdownToggle.removeClass(b),t.dropdownList.removeClass(y),t.overlay&&t.overlay.children().length&&(_.length?t.menu.insertAfter(_):t.menu.prependTo(t.parent),t.overlay.attr("style","").hide()),t.el.triggerHandler("w-close"),t.button.attr("aria-expanded","false")}}return c})},function(t,e,n){"use strict";var i=n(17);function r(t,e){var n=document.createEvent("CustomEvent");n.initCustomEvent(e,!0,!0,null),t.dispatchEvent(n)}var o=window.jQuery,a={},s={reset:function(t,e){i.triggers.reset(t,e)},intro:function(t,e){i.triggers.intro(t,e),r(e,"COMPONENT_ACTIVE")},outro:function(t,e){i.triggers.outro(t,e),r(e,"COMPONENT_INACTIVE")}};a.triggers={},a.types={INTRO:"w-ix-intro.w-ix",OUTRO:"w-ix-outro.w-ix"},o.extend(a.triggers,s),t.exports=a},function(t,e,n){"use strict";var i=window.jQuery,r={},o=[],a={reset:function(t,e){e.__wf_intro=null},intro:function(t,e){e.__wf_intro||(e.__wf_intro=!0,i(e).triggerHandler(r.types.INTRO))},outro:function(t,e){e.__wf_intro&&(e.__wf_intro=null,i(e).triggerHandler(r.types.OUTRO))}};r.triggers={},r.types={INTRO:"w-ix-intro.w-ix",OUTRO:"w-ix-outro.w-ix"},r.init=function(){for(var t=o.length,e=0;e<t;e++){var n=o[e];n[0](0,n[1])}o=[],i.extend(r.triggers,a)},r.async=function(){for(var t in a){var e=a[t];a.hasOwnProperty(t)&&(r.triggers[t]=function(t,n){o.push([e,n])})}},r.async(),t.exports=r}]);
