--- app/templates/base.html
@@ -1,75 +1,0 @@
-<!DOCTYPE html><!--  This site was created in Webflow. http://www.webflow.com  -->
-<!--  Last Published: Sat May 08 2021 11:33:04 GMT+0000 (Coordinated Universal Time)  -->
-<html lang="en" data-wf-page="6094f4c2f1b9885e5a5ef229" data-wf-site="6094f4c1f1b98891135ef228">
-
-<head>
-    <meta charset="utf-8">
-    {% if title %}
-    <title>{{ title }} | Dogecoin Is Awesome!</title>
-    {% else %}
-    <title>Dogecoin Is Awesome!</title>
-    {% endif %}
-    <meta name="author" content="Jacek Trociński">
-    <meta name="keywords" content="dogecoin, doge, how to, guide">
-    <meta content="Find everything you need to know about Dogecoin." name="description">
-    <meta content="Dogecoin Is Awesome!" property="og:title">
-    <meta content="Find everything you need to know about Dogecoin." property="og:description">
-    <meta content="https://dogecoinisawesome-dot-com.s3-us-west-2.amazonaws.com/open_graph_preview.jpeg"
-        property="og:image">
-    <meta content="Dogecoin Is Awesome!" property="twitter:title">
-    <meta content="Find everything you need to know about Dogecoin." property="twitter:description">
-    <meta content="https://dogecoinisawesome-dot-com.s3-us-west-2.amazonaws.com/open_graph_preview.jpeg"
-        property="twitter:image">
-    <meta property="og:type" content="website">
-    <meta content="summary_large_image" name="twitter:card">
-    <meta content="width=device-width, initial-scale=1" name="viewport">
-    <meta content="Webflow" name="generator">
-    <link href="css/normalize.css" rel="stylesheet" type="text/css">
-    <link href="css/webflow.css" rel="stylesheet" type="text/css">
-    <link href="css/dogecoinisawesome.webflow.css" rel="stylesheet" type="text/css">
-    <script src="https://ajax.googleapis.com/ajax/libs/webfont/1.6.26/webfont.js" type="text/javascript"></script>
-    <script type="text/javascript">WebFont.load({ google: { families: ["Mulish:regular,700"] } });</script>
-    <!-- [if lt IE 9]><script src="https://cdnjs.cloudflare.com/ajax/libs/html5shiv/3.7.3/html5shiv.min.js" type="text/javascript"></script><![endif] -->
-    <script
-        type="text/javascript">!function (o, c) { var n = c.documentElement, t = " w-mod-"; n.className += t + "js", ("ontouchstart" in o || o.DocumentTouch && c instanceof DocumentTouch) && (n.className += t + "touch") }(window, document);</script>
-    <link href="images/favicon.png" rel="shortcut icon" type="image/x-icon">
-    <link href="images/webclip.png" rel="apple-touch-icon">
-</head>
-
-<body class="body">
-    <div data-collapse="small" data-animation="default" data-duration="400" role="banner" class="navbar-2 w-nav">
-        <nav class="container-4 w-container">
-            <ul role="list" class="nav-grid-2 w-list-unstyled">
-                <li id="w-node-_74c1c4d6-2c99-458e-4548-7ed9839b7517-5a5ef229">
-                    <a href="/index" class="nav-logo-link-2"><img src="images/61081_DogecoinIsAwesome-com_logo_H_01.png"
-                            alt="DogecoinIsAwesome.com"
-                            sizes="(max-width: 767px) 82vw, (max-width: 991px) 57vw, 505.8125px"
-                            srcset="images/61081_DogecoinIsAwesome-com_logo_H_01-p-500.png 500w, images/61081_DogecoinIsAwesome-com_logo_H_01-p-800.png 800w, images/61081_DogecoinIsAwesome-com_logo_H_01-p-1080.png 1080w, images/61081_DogecoinIsAwesome-com_logo_H_01-p-1600.png 1600w, images/61081_DogecoinIsAwesome-com_logo_H_01-p-2000.png 2000w, images/61081_DogecoinIsAwesome-com_logo_H_01.png 2175w"
-                            class="nav-logo-2"></a>
-                </li>
-                <li>
-                    <div class="w-nav-button">
-                        <div class="w-icon-nav-menu"></div>
-                    </div>
-                    <nav role="navigation" class="nav-menu-2 w-nav-menu">
-                        <a href="/full-node" class="w-nav-link">Running a Full Node</a>
-                        <a href="https://discord.gg/MtHFyR2Z" class="w-nav-link">Discord</a>
-                    </nav>
-                </li>
-            </ul>
-        </nav>
-    </div>
-    {% block content %}{% endblock %}
-    <footer id="footer" class="footer">
-        <div class="w-container">
-            <div>Copyright © 2021 DogecoinIsAwesome.com. All rights reserved.</div>
-        </div>
-    </footer>
-    <script src="https://d3e54v103j8qbb.cloudfront.net/js/jquery-3.5.1.min.dc5e7f18c8.js?site=6094f4c1f1b98891135ef228"
-        type="text/javascript" integrity="sha256-9/aliU8dGd2tb6OSsuzixeV4y/faTqgFtohetphbbj0="
-        crossorigin="anonymous"></script>
-    <script src="js/webflow.js" type="text/javascript"></script>
-    <!-- [if lte IE 9]><script src="https://cdnjs.cloudflare.com/ajax/libs/placeholders/3.0.2/placeholders.min.js"></script><![endif] -->
-</body>
-
-</html>
