--- app/templates/full-node.html
@@ -1,12 +1,0 @@
-{% extends "base.html" %}
-
-{% block content %}
-<main id="main-content" class="main-content">
-    <div class="w-container">
-        <div class="w-richtext">
-            {{ markdown_html |safe }}
-        </div>
-    </div>
-</main>
-
-{% endblock %}
