--- app/templates/index.html
@@ -1,46 +1,0 @@
-{% extends "base.html" %}
-
-{% block content %}
-<section id="call-to-action" class="call-to-action">
-  <div class="centered-container w-container">
-    <h2 class="heading">Join the Discord Chat</h2><img src="images/doge.png" loading="lazy" alt="Doge">
-    <h6 class="heading-3">Join the Dogecoin Discord where we talk about Dogecoin and other things!</h6>
-    <a href="https://discord.gg/MtHFyR2Z" class="button-hero w-button">Discord</a>
-  </div>
-</section>
-<main id="main-content" class="main-content">
-  <div class="w-container">
-    <div class="w-richtext">
-      <h1>Why is Dogecoin Awesome?</h1>
-      <p>Dogecoin uses peer-to-peer technology to operate with no central authority or banks; managing transactions
-        and the issuing of Dogecoins is carried out collectively by the network. Dogecoin isopen-source; its design is
-        public, nobody owns or controls Dogecoin and everyone can take part.</p>
-      <p>‍</p>
-      <h3>A Little History</h3>
-      <p>The first few weeks after Dogecoin was released were incredible, it went from a being a silly joke to
-        something worth something to people and a community was developing fast, with lots of people putting up
-        services and infrastructure around it.</p>
-      <p>One of those things was an online wallet - instead of having to download the clunky and slow main client
-        someone had provided a service to hold people&#x27;s Dogecoin for them. Online wallets have a history of being
-        sketchy, and sure enough, on Christmas, the wallet was hacked, and 21 million Dogecoins were stolen in the
-        process.</p>
-      <p>This type of story was all too common in the cryptocurrency world, but what happened next was not. Some
-        members of the community got together and started a Dogecoin fund to &quot;Save Dogemas&quot; - to try to
-        return money to those who had lost it. The Dogecoin community donated millions of coins to help those who had
-        been victims of the bad actor - and that spirit didn&#x27;t stop, the community would donate Dogecoins for
-        both good causes, such as helping build water wells in Africa, helping connect service dogs and kids, and
-        silly and amusing causes, such as helping send the Jamaican Bobsled team to the Olympics, or sponsoring a
-        Nascar driver.</p>
-      <p>‍</p>
-      <figure class="w-richtext-align-center w-richtext-figure-type-image">
-        <div><img src="images/dogecoin-nascar.jpg" loading="lazy" alt="Dogecoin Nascar"></div>
-      </figure>
-      <p>‍</p>
-      <h3>What’s the True Value of Dogecoin?</h3>
-      <p>Joy, kindness, learning, giving, empathy, fun, community, inspiration, creativity, generosity, silliness,
-        absurdity. These are the things that make Dogecoin what it is and when the community embodies these things,
-        that&#x27;s True Value.</p>
-    </div>
-  </div>
-</main>
-{% endblock %}
