+++ requirements.txt
@@ -9,7 +9,7 @@ Markdown==3.3.4
 MarkupSafe==1.1.1
 mypy-extensions==0.4.3
 pathspec==0.8.1
-python-dotenv==0.17.1
+python-dotenv==0.16.0
 regex==2020.11.13
 toml==0.10.2
 typed-ast==1.4.2
