+++ static/css/dogecoinisawesome.css
@@ -1,0 +1,84 @@
+body,
+html {
+    background: #f1c40f url("/static/img/bg.jpeg") no-repeat center center fixed;
+    background-size: cover;
+    font-family: "Comic Neue", cursive;
+    padding: 0;
+    margin: 0;
+}
+
+body {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    height: 100vh;
+    color: #586069;
+}
+
+main {
+    box-sizing: border-box;
+    min-height: 100%;
+    max-width: 100%;
+    text-align: center;
+}
+
+h1,
+h2,
+h3 {
+    margin: 0;
+    padding: 0;
+    line-height: 1.2;
+    color: #24292e;
+}
+
+h1 {
+    font-size: 1.4em;
+}
+
+h2 {
+    font-size: 1.2em;
+}
+
+a {
+    color: #0366d6;
+    text-decoration: none;
+}
+
+a:hover {
+    text-decoration: underline;
+}
+
+img {
+    max-width: 100%;
+}
+
+code,
+pre {
+    font-family: "Roboto Mono", monospace;
+    padding: 0.5em;
+    box-sizing: border-box;
+    line-height: 1.4;
+}
+
+pre {
+    overflow-x: auto;
+    background-color: #f6f8fa;
+}
+
+nav,
+article {
+    background-color: #fff;
+    border: 0.05em solid #24292e;
+    border-radius: 4px;
+    padding: 1em;
+    box-sizing: border-box;
+    width: 50em;
+    max-width: calc(100% - 1em);
+    text-align: left;
+    margin: 0.5em;
+}
+
+footer {
+    padding: 1em;
+    text-align: center;
+}
