+++ static/img/full-node/active-connections.png
