+++ static/img/full-node/debug-inbound-connections.png
