+++ templates/application-error.html
@@ -1,0 +1,18 @@
+{% include "head.html" %}
+
+<body>
+    <main>
+        <nav>
+            <h2><a href="/">Home</a></h2>
+        </nav>
+
+        <article>
+            <h1>Application error</h1>
+            <p>An error occurred in the application and your page could not be served.</p>
+        </article>
+
+        {% include "footer.html" %}
+    </main>
+</body>
+
+</html>
