+++ templates/footer.html
@@ -1,0 +1,1 @@
+<footer>Donate Dogecoin to keep this site running! Dogecoin address: DDNhF6qpFPHM1KxLUWJp3WPa57QCX1Jt2F</footer>
