+++ templates/full-node.html
@@ -1,0 +1,18 @@
+{% include "head.html" %}
+
+<body>
+    <main>
+        <nav>
+            <h2><a href="/">Home</a></h2>
+        </nav>
+
+        <article>
+            {{ markdown_html |safe }}
+        </article>
+
+        {% include "footer.html" %}
+    </main>
+
+</body>
+
+</html>
