+++ templates/index.html
@@ -1,0 +1,62 @@
+{% include "head.html" %}
+
+<body>
+    <main>
+        <img src="/static/img/doge.png" alt="Dogecoin Is Awesome." />
+
+        <nav>
+            <h2><a href="/full-node">Running a Full Node</a></h2>
+        </nav>
+
+        <article>
+            <h1>Why is Dogecoin Awesome?</h1>
+            <p>
+                Dogecoin uses peer-to-peer technology to operate with no central authority or banks; managing
+                transactions and the issuing of Dogecoins is carried out collectively by the network. Dogecoin is
+                open-source; its design is public, nobody owns or controls Dogecoin and everyone can take part.
+            </p>
+            <h2>A Little History<sup>1</sup></h2>
+            <p>
+                The first few weeks after Dogecoin was released were incredible, it went from a being a silly joke to
+                something worth something to people and a community was developing fast, with lots of people putting up
+                services and infrastructure around it.
+            </p>
+            <p>
+                One of those things was an online wallet - instead of having to download the clunky and slow main client
+                someone had provided a service to hold people's Dogecoin for them. Online wallets have a history of
+                being sketchy, and sure enough, on Christmas, the wallet was hacked, and 21 million Dogecoins were
+                stolen in the process.
+            </p>
+            <p>
+                This type of story was all too common in the cryptocurrency world, but what happened next was not. Some
+                members of the community got together and started a Dogecoin fund to "Save Dogemas" - to try to return
+                money to those who had lost it. The Dogecoin community donated millions of coins to help those who had
+                been victims of the bad actor - and that spirit didn't stop, the community would donate Dogecoins for
+                both good causes, such as helping build water wells in Africa, helping connect service dogs and kids,
+                and silly and amusing causes, such as helping send the Jamaican Bobsled team to the Olympics, or
+                sponsoring a Nascar driver.
+            </p>
+            <p><img src="/static/img/dogecoin-nascar.jpg" alt="Dogecoin Is Awesome." /></p>
+            <h2>What’s the True Value of Dogecoin?<sup>1</sup></h2>
+            <p>
+                Joy, kindness, learning, giving, empathy, fun, community, inspiration, creativity, generosity,
+                silliness, absurdity. These are the things that make Dogecoin what it is and when the community embodies
+                these things, that's True Value.
+            </p>
+            <h3>Sources</h3>
+            <ol>
+                <li>
+                    <a href="https://www.reddit.com/r/dogecoin/comments/lfl5iz/true_value_an_open_letter_from_billy_markus/"
+                        target="_blank" rel="noopener noreferrer">
+                        True value - an open letter from Billy Markus, co-founder and original creator of Dogecoin
+                    </a>
+                </li>
+            </ol>
+        </article>
+
+        {% include "footer.html" %}
+    </main>
+
+</body>
+
+</html>
